+++ imod/typing/__init__.py
@@ -2,6 +2,7 @@
 Module to define type aliases.
 """
 
+from datetime import datetime
 from typing import TYPE_CHECKING, Literal, TypeAlias, TypedDict, TypeVar, Union
 
 import numpy as np
@@ -17,6 +18,7 @@ ScalarAsDataset: TypeAlias = Union[xr.Dataset, xu.UgridDataset]
 UnstructuredData: TypeAlias = Union[xu.UgridDataset, xu.UgridDataArray]
 FloatArray: TypeAlias = NDArray[np.floating]
 IntArray: TypeAlias = NDArray[np.int_]
+StressPeriodTimesType: TypeAlias = list[datetime] | Literal["steady-state"]
 
 
 class SelSettingsType(TypedDict, total=False):